big commit: refactor and add channel check
change util locations:
* interfaces moved to shared/interactions
* removed getAttachment, moved getStream to voice/utils
