+++ src/commands/slash/voice/clear.ts
@@ -1,6 +1,7 @@
 import { ApplicationCommandData } from "discord.js";
-import { cmdEvent } from '@shared/interfaces'
+import { cmdEvent } from '@interactions/interfaces'
 import { subscriptions } from "@shared/voice/subscription";
+import { userInCorrectChannel } from "@interactions/utils";
 
 export const data: ApplicationCommandData = {
     name: "clear",
@@ -9,14 +10,17 @@ export const data: ApplicationCommandData = {
 
 export async function run(e: cmdEvent) {
     if (!e.interaction.guildId) return;
-    const subscription = subscriptions.get(e.interaction.guildId);
 
-    if (subscription) {
-        subscription.stop();
-        await e.interaction.reply(`Cleared the queue!`);
-        subscription.queueLock = false;
-    }
-    else {
+    const subscription = subscriptions.get(e.interaction.guildId)
+    
+    if (!subscription) {
         await e.interaction.reply('Not in a voice channel!');
+        return;
     }
+    
+    if (!userInCorrectChannel(e, subscription)) return;
+
+    subscription.stop();
+    await e.interaction.reply(`Cleared the queue!`);
+    subscription.queueLock = false;
 }
