+++ src/commands/slash/voice/leave.ts
@@ -1,7 +1,7 @@
 import { ApplicationCommandData } from "discord.js";
-import { cmdEvent } from '@shared/interfaces'
-
+import { cmdEvent } from '@interactions/interfaces'
 import { subscriptions } from "@shared/voice/subscription";
+import { userInCorrectChannel } from "@interactions/utils";
 
 export const data: ApplicationCommandData = {
     name: "leave",
@@ -11,14 +11,16 @@ export const data: ApplicationCommandData = {
 export async function run(e: cmdEvent) {
     if (!e.interaction.guildId) return;
 
-    const subscription = subscriptions.get(e.interaction.guildId);
+    const subscription = subscriptions.get(e.interaction.guildId)
 
-    if (subscription) {
-        subscription.voiceConnection.destroy();
-        subscriptions.delete(e.interaction.guildId)
-        await e.interaction.reply('Left channel!')
-    }
-    else {
+    if (!subscription) {
         await e.interaction.reply('Not in a voice channel!');
+        return;
     }
+
+    if (!userInCorrectChannel(e, subscription)) return;
+
+    subscription.voiceConnection.destroy();
+    subscriptions.delete(e.interaction.guildId)
+    await e.interaction.reply('Left channel!')
 }
