+++ src/commands/slash/voice/skip.ts
@@ -1,6 +1,7 @@
 import { ApplicationCommandData } from "discord.js";
-import { cmdEvent } from "@shared/interfaces";
+import { cmdEvent } from "@interactions/interfaces";
 import { subscriptions } from "@voice/subscription";
+import { userInCorrectChannel } from "@interactions/utils";
 
 export const data: ApplicationCommandData = {
     name: 'skip',
@@ -9,16 +10,19 @@ export const data: ApplicationCommandData = {
 
 export async function run(e: cmdEvent) {
     if (!e.interaction.guildId) return;
-    
+
     const subscription = subscriptions.get(e.interaction.guildId);
 
-    if (subscription) {
-        // Calling .stop() on an AudioPlayer causes it to transition into the Idle state. Because of a state transition
-        // listener defined in music/subscription.ts, transitions into the Idle state mean the next track from the queue
-        // will be loaded and played.
-        subscription.audioPlayer.stop();
-        await e.interaction.reply('Skipped song!');
-    } else {
-        await e.interaction.reply('Not playing in this server!');
+    if (!subscription) {
+        await e.interaction.reply('Not in a voice channel!');
+        return;
     }
+
+    if (!userInCorrectChannel(e, subscription)) return;
+    
+    // Calling .stop() on an AudioPlayer causes it to transition into the Idle state. Because of a state transition
+    // listener defined in music/subscription.ts, transitions into the Idle state mean the next track from the queue
+    // will be loaded and played.
+    subscription.audioPlayer.stop();
+    await e.interaction.reply('Skipped song!');
 }
