+++ src/shared/embed.ts
@@ -2,10 +2,10 @@
  * This file can be used to store custom embeds such as the one below.
  */
 
-import Discord from 'discord.js'
+import { Message, MessageEmbed } from 'discord.js'
 
-export function AutoreplyEmbedGen(name: string, imgurl: string, msg: Discord.Message, link?: string) {
-    const embed = new Discord.MessageEmbed()
+export function AutoreplyEmbedGen(name: string, imgurl: string, msg: Message, link?: string) {
+    const embed = new MessageEmbed()
         .setImage(imgurl)
         .setFooter(name + " from " + msg.author.tag); // Can put a link here if we want (idk what)
     if (link) embed.setURL(link);
