+++ src/shared/interactions/interfaces.ts
