+++ src/shared/voice/utils.ts
@@ -89,3 +89,15 @@ export async function connectToChannel(channel: VoiceChannel) {
 		throw error;
 	}
 }
+
+import https from 'https';
+import { PassThrough } from 'stream';
+
+export async function getStream(url: string) {
+    let stream = new PassThrough();
+
+    https.get(url, (res) => res.pipe(stream))
+
+    //await new Promise(fullfill => stream.on('end',fullfill))
+    return stream;
+}
